move error fallback to own component
